Combine 2 parital upate file into 1
